refactor: change checkbox component path
