+++ src/app/(user)/register/_components/policy/checkbox.tsx
